+++ issuer-application/src/test/java/ch/admin/bj/swiyu/issuer/service/DemonstratingProofOfPossessionIT.java
@@ -58,7 +58,7 @@ class DemonstratingProofOfPossessionIT {
     private UUID accessToken;
     @Autowired
     private CredentialManagementRepository credentialManagementRepository;
-    
+
     public static Stream<String> faultyNonceSource() {
         var nonceSecret = IssuerSecret.builder().id(UUID.randomUUID()).build();
         return Stream.of(
@@ -81,7 +81,7 @@ class DemonstratingProofOfPossessionIT {
     void setUp() {
         accessToken = UUID.randomUUID();
 
-        testCredentialOffer = CredentialOffer.builder()
+        var createTestCredentialOffer = CredentialOffer.builder()
                 .id(UUID.randomUUID())
                 .preAuthorizedCode(UUID.randomUUID())
                 .credentialStatus(CredentialOfferStatusType.OFFERED)
@@ -98,13 +98,17 @@ class DemonstratingProofOfPossessionIT {
 
         var mgmt = credentialManagementRepository.save(credentialManagement);
 
-        testCredentialOffer.setCredentialManagement(credentialManagement);
+        createTestCredentialOffer.setCredentialManagement(credentialManagement);
 
-        credentialOfferRepository.save(testCredentialOffer);
+        credentialOfferRepository.save(createTestCredentialOffer);
         credentialManagement.setCredentialOffers(
-                Set.of(testCredentialOffer)
+                Set.of(createTestCredentialOffer)
         );
         credentialManagementRepository.save(mgmt);
+
+        // Reload Object to ensure that generated values are all present
+        // notably audit data is without this step null and will cause DataIntegrityViolationException
+        testCredentialOffer = credentialOfferRepository.findById(createTestCredentialOffer.getId()).get();
 
         dpopKey = assertDoesNotThrow(() -> new ECKeyGenerator(Curve.P_256).keyID("test-key-1").keyUse(KeyUse.SIGNATURE).generate());
     }
@@ -157,7 +161,7 @@ class DemonstratingProofOfPossessionIT {
         var tokenRequest = mockTokenHttpRequest(baseTestHttpRequest);
         var registrationDPoP = getDPoPJWT(tokenRequest, faultyNonce, null);
         String preAuthCode = testCredentialOffer.getPreAuthorizedCode().toString();
-        assertThrows(DemonstratingProofOfPossessionException.class, () -> demonstratingProofOfPossessionService.registerDpop(preAuthCode, registrationDPoP, tokenRequest));
+        assertThrows(DemonstratingProofOfPossessionException.class, () -> demonstratingProofOfPossessionService.registerDpop(testCredentialOffer, registrationDPoP, tokenRequest));
     }
 
     private HttpRequest createMockRequest() {
@@ -167,7 +171,8 @@ class DemonstratingProofOfPossessionIT {
     private void registerDPoP(HttpRequest baseTestHttpRequest, String nonce) {
         var tokenRequest = mockTokenHttpRequest(baseTestHttpRequest);
         var registrationDPoP = getDPoPJWT(tokenRequest, nonce, null);
-        assertDoesNotThrow(() -> demonstratingProofOfPossessionService.registerDpop(testCredentialOffer.getPreAuthorizedCode().toString(), registrationDPoP, tokenRequest));
+
+        assertDoesNotThrow(() -> demonstratingProofOfPossessionService.registerDpop(testCredentialOffer, registrationDPoP, tokenRequest));
     }
 
     private HttpRequest mockTokenHttpRequest(HttpRequest baseTestHttpRequest) {