+++ issuer-service/src/main/java/ch/admin/bj/swiyu/issuer/common/config/ApplicationProperties.java
@@ -147,6 +147,14 @@ public class ApplicationProperties {
     @Min(1)
     private Integer maxDecompressedPayloadLength;
 
+    /**
+     * Limit how many retries a holder has to provide the correct Transaction Code when requesting the token.
+     * Note that the total number of attempts a wallet may make is retries + 1 for the initial try
+     */
+    @NotNull
+    @Min(0)
+    private Integer txCodeRetries;
+
     @PostConstruct
     public void init() {
         try {