+++ issuer-service/src/main/java/ch/admin/bj/swiyu/issuer/service/SdJwtCredential.java
@@ -20,7 +20,7 @@ import ch.admin.bj.swiyu.sdjwtbuilder.exception.SdJwtBuilderException;
 import ch.admin.bj.swiyu.sdjwtbuilder.SdJwtVcBuilder.CreatedSdJwtVc;
 import ch.admin.bj.swiyu.statuslist.dto.TokenStatusListReferenceDto.TokenStatusListStatusListReference;
 
-import com.nimbusds.jose.*;
+import com.nimbusds.jose.JWSSigner;
 import com.nimbusds.jose.jwk.JWK;
 import jakarta.annotation.Nullable;
 import lombok.extern.slf4j.Slf4j;